requirement.txt
requests
pytest
allure-pytest
python-dotenv
jsonschema